requirements.txt
mysqlclient==1.3.12
PyYAML==3.13
